Refactor test structure (#50)
The tests had been setup with all the cases in one location, then each database having its own file to arrange/act/assert based on those cases. A few things precipitated an inversion where now the test cases are split up among files and the test logic is shared between databases.

1. There's no sqlite so it is easier to make one database call
1. We added a 4th type of test with history and are about to add many many more for the transformation rework
1. The test setup was getting hard to even understand and modify
1. The end to end tests would be nice to have running on postgres

As part of this refactor the slow-running integration tests are now filterable based on the Integration name. I've also mad the end to end tests run against both duckdb and postgres. There's also lots of misc database and sql cleanup pieces in tests where I wasn't using the libraries in the best ways before.
